+++ Apptentive/Apptentive/Message Center/Views/ApptentiveMessageCenterProfileView.h
@@ -10,22 +10,15 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-typedef NS_ENUM(NSInteger, ATMessageCenterProfileMode) {
-	ATMessageCenterProfileModeCompact = 1,
-	ATMessageCenterProfileModeFull
-};
-
 
 @interface ApptentiveMessageCenterProfileView : UIView
 
-@property (assign, nonatomic) ATMessageCenterProfileMode mode;
 @property (weak, nonatomic) IBOutlet UIView *containerView;
 @property (weak, nonatomic) IBOutlet UILabel *titleLabel;
 @property (weak, nonatomic) IBOutlet UITextField *nameField;
 @property (weak, nonatomic) IBOutlet UITextField *emailField;
 @property (weak, nonatomic) IBOutlet UIButton *skipButton;
 @property (weak, nonatomic) IBOutlet UIButton *saveButton;
-@property (weak, nonatomic) IBOutlet UILabel *requiredLabel;
 @property (strong, nonatomic) UIColor *borderColor;
 
 @end